# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Ubiq CMS Member Web Service
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /member:
    x-swagger-router-controller: ubiq_cms_member
    get:
      description: Returns Member Details, Status and Redemption Balance
      operationId: memberDetails
      parameters:
        - name: customerID
          in: query
          description: The customer ID of the member
          required: true
          type: string
        - name: posTerminalID
          in : query
          description : The POS Terminal making the query
          required: true
          type: string
      responses:
        "200" :
          description: Success
          schema:
            # A Pointer to a Response Definition
            $ref: "#/definitions/MemberDetailResponse"
        "404" :
          description: Error
          schema:
            $ref: "#/definitions/MemberDetailErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/MemberDetailErrorResponse"
  /billing:
    x-swagger-router-controller: ubiq_cms_billing
    post:
      description: Posts a billing transaction and deduct Redemptions
      operationId: postTransaction
      parameters:
        - name: transactionHeader
          in: body
          description: The Transaction Header
          required: true
          schema:
            $ref: "#/definitions/PostBillingTransaction"
      responses:
        "200" :
          description: Success
          schema:
            $ref: "#/definitions/PostBillingResponse"
        "404" :
          description: Error
          schema:
            $ref: "#/definitions/PostBillingErrorResponse"
        "400" :
          description: Error
          schema:
            $ref: "#/definitions/PostBillingErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  MemberDetailResponse:
    required:
      - customerID
      - posTerminalID
      - membershipNo
      - memberName
      - memberStatus
      - redemptionBalance
    properties:
      customerID:
        type: string
      posTerminalID:
        type: string
      membershipNo:
        type: string
      memberName:
        type: string
      memberStatus:
        type: string
      redemptionBalance:
        type: number
        format: double
  MemberDetailErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  PostBillingTransaction:
    required:
      - posTerminalID
      - customerID
      - billNo
      - outletCode
      - membershipNo
      - totalAmount
      - serviceCharge
      - taxAmount
    properties:
      posTerminalID:
        type: string
      customerID:
        type: string
      billNo:
        type: string
      outletCode:
        type: string
      membershipNo:
        type: string
      transactionDate:
        type: string
      canceled:
        type: string
      totalAmount:
        type: number
      serviceCharge:
        type: number
      taxAmount:
        type: number
      details:
        type: array
        items:
          $ref: "#/definitions/PostBillingTransactionDetails"
      payments:
        type: array
        items:
          $ref: "#/definitions/PostBillingPaymentDetails"
  PostBillingTransactionDetails:
    required:
      - billNo
      - itemCode
      - price
      - qty
      - totalAmount
    properties:
      billNo:
        type: string
      itemCode:
        type: string
      price:
        type: number
      qty:
        type: number
      totalAmount:
        type: number
  PostBillingPaymentDetails:
    required:
      - billNo
      - seqNo
      - paymentMode
      - amount
    properties:
      billNo:
        type: string
      seqNo:
        type: string
      paymentMode:
        type: string
      amount:
        type: number
  PostBillingErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  PostBillingResponse:
    required:
      - billNo
      - customerID
      - posTerminalID
      - redemptionBalance
    properties:
      billNo:
        type: string
      customerID:
        type: string
      posTerminalID:
        type: string
      redemptionBalance:
        type: number
        format: dobule